ai-services/whatsapp-analysis/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
textblob==0.17.1
python-multipart==0.0.6
pydantic==2.5.0
Pillow>=10.0.0
pytesseract>=0.3.10
easyocr>=1.7.0
opencv-python>=4.8.0

